+++ crates/tau-runtime/Cargo.toml
@@ -10,6 +10,7 @@ serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
 tau-ai = { path = "../tau-ai" }
+tau-agent-core = { path = "../tau-agent-core" }
 tau-core = { path = "../tau-core" }
 
 [dev-dependencies]
